refactor: move preferencesPlugin
