+++ plugins/preferences/package.json
@@ -93,6 +93,7 @@
     "typescript": "^5.0.0"
   },
   "peerDependencies": {
+    "@just-web/app": "workspace:^",
     "@just-web/commands": "workspace:^",
     "@just-web/keyboard": "workspace:^",
     "@just-web/log": "workspace:^",
