+++ plugins/preferences/ts/index.ts
@@ -1,2 +1,3 @@
 export * from './preferences.js'
-export { default } from './preferences.js'
+export { preferencesPlugin as default } from './preferences_plugin.js'
+export type { PreferencesContext } from './preferences_plugin.js'
